implement typograph for text when not editing, updated file names for conformity
