+++ client/features/output-visualization-page/components/Flow.tsx
@@ -13,7 +13,7 @@ import ReactFlow, {
 } from 'reactflow';
 import 'reactflow/dist/style.css';
 import LinktaFlowEdge from '@features/output-visualization-page/components/LinktaFlowEdge';
-import LinktaNode from '@features/output-visualization-page/components/LinktaNode';
+import LinktaFlowNode from '@features/output-visualization-page/components/LinktaFlowNode';
 import ConnectionLine from '@features/output-visualization-page/components/ConnectionLine';
 import useFetchLinktaFlow from '@hooks/useFetchLinktaFlow';
 import useLinktaFlowStore from '@stores/LinktaFlowStore';
@@ -22,7 +22,7 @@ import { useNavigate } from 'react-router-dom';
 import Loader from '@/components/common/Loader';
 import UndoAndRedo from '@features/output-visualization-page/components/UndoAndRedo';
 
-const nodeTypes = { linktaNode: LinktaNode };
+const nodeTypes = { linktaNode: LinktaFlowNode };
 const edgeTypes = { linktaEdge: LinktaFlowEdge };
 
 const rfStyle = {
